+++ workspaces/marketplace/plugins/marketplace/src/components/MarketplacePackageInstallContent.tsx
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import React, { useState } from 'react';
+import React, { useEffect, useState } from 'react';
 
 import { ErrorPage, Progress } from '@backstage/core-components';
 import {
@@ -142,6 +142,17 @@ export const MarketplacePackageInstallContent = ({
 }: {
   pkg: MarketplacePackage;
 }) => {
+  const [hasGlobalHeader, setHasGlobalHeader] = useState(false);
+
+  useEffect(() => {
+    const header = document.querySelector('nav#global-header');
+    setHasGlobalHeader(Boolean(header));
+  }, []);
+
+  const dynamicHeight = hasGlobalHeader
+    ? 'calc(100vh - 220px)'
+    : 'calc(100vh - 160px)';
+
   const codeEditor = useCodeEditor();
   const params = useRouteRefParams(pluginInstallRouteRef);
 
@@ -183,7 +194,7 @@ export const MarketplacePackageInstallContent = ({
   return (
     <Box
       sx={{
-        height: 'calc(100vh - 160px)',
+        height: dynamicHeight,
         width: '100%',
         display: 'flex',
         flexDirection: 'column',
@@ -320,7 +331,7 @@ export const MarketplacePackageInstallContent = ({
         sx={{
           mt: 4,
           flexShrink: 0,
-          backgroundColor: 'background.paper',
+          backgroundColor: 'inherit',
         }}
       >
         <Button variant="contained" color="primary" disabled>