+++ workspaces/marketplace/plugins/marketplace/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @red-hat-developer-hub/backstage-plugin-marketplace
 
+## 0.5.3
+
+### Patch Changes
+
+- 46b3262: add missing grid spacing when loaded as dynamic plugin
+
+## 0.5.2
+
+### Patch Changes
+
+- ff849f7: Minor UI issues fixed
+
 ## 0.5.1
 
 ### Patch Changes